+++ src/redux/reducers.js
@@ -2033,12 +2033,14 @@ export function betaChannel(
   return state;
 }
 
-export function appUpdated(state = { version: null, updated: false }, action) {
+export function appUpdated(state = { version: null, initialVersion: null }, action) {
   if (action.type === ActionTypes.APP_VERSION) {
     const version = action.payload;
-    const updated = state.version !== null && state.version !== version;
-    if (state.version === null || updated) {
-      return { version, updated };
+    if (state.initialVersion === null) {
+      return { version, initialVersion: version };
+    }
+    if (version !== state.initialVersion) {
+      return { ...state, version };
     }
   }
   return state;